Extract VotingRound to separate file
This is part of a major code refactoring to
improve the project's structure.

Also:
- files added for constants and printing
- some unnecessary includes removed
