+++ src/CMakeLists.txt
@@ -1,9 +1,14 @@
 cmake_minimum_required(VERSION 3.15)
 
 add_executable(${PROJECT_NAME}
+	constants.h
 	functions.cpp
 	functions.h
 	pairwise-ranking.cpp
+	print.cpp
+	print.h
+	voting_round.cpp
+	voting_round.h
 )
 
 add_subdirectory(test)
