+++ src/pairwise-ranking.cpp
@@ -15,7 +15,9 @@
 #include <type_traits>
 #include <vector>
 
+#include "constants.h"
 #include "functions.h"
+#include "print.h"
 
 void programLoop() {
 	std::optional<VotingRound> voting_round{};
