+++ src/print.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include <string>
+
+void print(std::string const& str, bool add_newline = true);
+void printError(std::string const& err);
