+++ src/test/test_generate_new_voting_round.cpp
@@ -1,3 +1,4 @@
+#include "constants.h"
 #include "functions.h"
 #include "testing.h"
