+++ src/test/test_prune_votes.cpp
@@ -1,3 +1,4 @@
+#include "constants.h"
 #include "functions.h"
 #include "testing.h"
